finished monitoring service
