+++ activity-service/build.gradle
@@ -37,7 +37,6 @@ dependencies {
     implementation(libs.grpc)
     implementation(libs.protobuf)
     implementation(libs.grpcStub)
-//    implementation(libs.grpcNettyShaded)
     runtimeOnly(libsRuntimeOnly.postgresql)
 
     testImplementation(libsTest.springbootStarterTest)
