+++ activity-service/src/main/java/com/example/services/ActivityService.java
@@ -73,7 +73,7 @@ public class ActivityService {
     }
 
     private void sendHeartBeat() {
-        scheduler.scheduleAtFixedRate(this::sendHeartBeatMessage, 0, 10000, TimeUnit.MILLISECONDS);
+        scheduler.scheduleAtFixedRate(this::sendHeartBeatMessage, 10, 10, TimeUnit.SECONDS);
     }
 
     private void sendHeartBeatMessage() {
