+++ monitoring-service/src/main/resources/application.properties
@@ -17,7 +17,7 @@ spring.jpa.show-sql = false
 spring.jpa.hibernate.naming.implicit-strategy = org.hibernate.boot.model.naming.ImplicitNamingStrategyJpaCompliantImpl
 spring.jpa.properties.hibernate.format_sql=true
 
-grpc.client.sweetcity.address=static://localhost:9095
-grpc.client.sweetcity.enable-keep-alive=true
-grpc.client.sweetcity.keep-alive-without-calls=true
-grpc.client.sweetcity.negotiationType=plaintext
+#grpc.client.sweetcity.address=static://localhost:9095
+#grpc.client.sweetcity.enable-keep-alive=true
+#grpc.client.sweetcity.keep-alive-without-calls=true
+#grpc.client.sweetcity.negotiationType=plaintext
