+++ search-service/build.gradle
@@ -18,7 +18,7 @@ repositories {
 
 dependencies {
     implementation project(":activity-service")
-//    implementation project(':monitoring-service')
+    implementation project(':monitoring-service')
     implementation(libs.springbootStarter)
     implementation(libs.springWeb)
     implementation(libs.springbootStarterWeb)
@@ -36,7 +36,6 @@ dependencies {
     implementation(libs.grpc)
     implementation(libs.protobuf)
     implementation(libs.grpcStub)
-//    implementation(libs.grpcNettyShaded)
 
     runtimeOnly(libsRuntimeOnly.postgresql)
