+++ search-service/src/main/resources/application.properties
@@ -1,6 +1,6 @@
 server.port=8081
 spring.application.name=search-service
-
+grpc.server.port=9091
 # Details for datasource
 spring.datasource.driver-class-name=org.postgresql.Driver
 spring.datasource.username=postgres
@@ -20,7 +20,7 @@ amazon.dynamodb.accesskey=test
 amazon.dynamodb.secretkey=test
 amazon.dynamodb.region=us-east-1
 
-#amazon.s3.endpointUrl=s3://sweet-city-storage
+
 amazon.s3.endpointUrl=http://localhost:4566
 amazon.s3.accesskey=test
 amazon.s3.secretkey=test
@@ -37,3 +37,7 @@ grpc.client.search.address=static://localhost:9095
 grpc.client.search.enable-keep-alive=true
 grpc.client.search.keep-alive-without-calls=true
 grpc.client.search.negotiationType=plaintext
+grpc.client.sweetcity.address=static://localhost:9095
+grpc.client.sweetcity.enable-keep-alive=true
+grpc.client.sweetcity.keep-alive-without-calls=true
+grpc.client.sweetcity.negotiationType=plaintext
